Add files via upload
The testbench was slightly modified to prevent the $monitor function from printing the result while the output signal is still unstable. 
The $monitor function was replaced with $display that appears before the signals are loaded.
This configuration makes the result of the last signal not appear, but as I used a long list of test vectors (100+), I take it to not be harmful.